api/requirements.txt
# FastAPI Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Client & External APIs
httpx>=0.25.0

# Data Validation & Serialization
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Environment & Configuration
python-dotenv>=1.0.0

# Date & Time Handling
python-dateutil>=2.8.2
pytz>=2023.3

# Existing dependencies from original project
notion-client>=2.2.1
anthropic>=0.7.8